+++ src/instruments/src/EFB/Settings/Pages/RealismPage.tsx
@@ -19,6 +19,8 @@ export const RealismPage = () => {
     const [dmcSelfTestTime, setDmcSelfTestTime] = usePersistentProperty('CONFIG_SELF_TEST_TIME', '12');
     const [mcduInput, setMcduInput] = usePersistentProperty('MCDU_KB_INPUT', 'DISABLED');
     const [mcduTimeout, setMcduTimeout] = usePersistentProperty('CONFIG_MCDU_KB_TIMEOUT', '60');
+    const [pauseAtTod, setPauseAtTod] = usePersistentProperty('PAUSE_AT_TOD', 'DISABLED');
+    const [todOffset, setTodOffset] = usePersistentNumberProperty('PAUSE_AT_TOD_DISTANCE', 10);
     const [boardingRate, setBoardingRate] = usePersistentProperty('CONFIG_BOARDING_RATE', 'REAL');
     const [realisticTiller, setRealisticTiller] = usePersistentNumberProperty('REALISTIC_TILLER_ENABLED', 0);
     const [homeCockpit, setHomeCockpit] = usePersistentProperty('HOME_COCKPIT_ENABLED', '0');
@@ -114,6 +116,28 @@ export const RealismPage = () => {
                             onChange={(event) => {
                                 if (!Number.isNaN(event) && parseInt(event) >= 5 && parseInt(event) <= 120) {
                                     setMcduTimeout(event.trim());
+                                }
+                            }}
+                        />
+                    </SettingItem>
+                )}
+            </SettingGroup>
+
+            <SettingGroup>
+                <SettingItem name={t('Settings.Realism.PauseAtTod')} unrealistic groupType="parent">
+                    <Toggle value={pauseAtTod === 'ENABLED'} onToggle={(value) => setPauseAtTod(value ? 'ENABLED' : 'DISABLED')} />
+                </SettingItem>
+                {pauseAtTod === 'ENABLED' && (
+                    <SettingItem name={t('Settings.Realism.PauseAtTodDistance')} groupType="sub">
+                        <SimpleInput
+                            className="text-center w-30"
+                            value={todOffset}
+                            min={0}
+                            max={50.0}
+                            disabled={(pauseAtTod !== 'ENABLED')}
+                            onChange={(event) => {
+                                if (!Number.isNaN(event) && parseInt(event) >= 0 && parseInt(event) <= 50.0) {
+                                    setTodOffset(parseFloat(event.trim()));
                                 }
                             }}
                         />